Feature/refactor fdr (#251)
* remove the use of macros for critical loops, easier to debug
removed switch, merged get_conf_stride functions into 1

* remove the use of macros for critical loops, easier to debug
removed switch, merged get_conf_stride functions into 1
split FDR implementations into arch specific files (same for now)
